move ontology api handling to api namespace.
