+++ spec/controllers/ontologies_controller_spec.rb
@@ -90,30 +90,6 @@ describe OntologiesController do
     context 'on GET to show' do
       before { OntologyMember::Symbol.delete_all }
 
-      context 'with format json', api_specification: true do
-        let(:ontology_schema) { schema_for('ontology') }
-
-        before do
-          get :show,
-            repository_id: repository.to_param,
-            format:        :json,
-            id:            ontology.to_param
-        end
-
-        it { should respond_with :success }
-
-        it 'respond with json content type' do
-          expect(response.content_type.to_s).to eq('application/json')
-        end
-
-        it 'should return a representation that validates against the schema' do
-          VCR.use_cassette 'api/json-schemata/ontology' do
-            expect(response.body).to match_json_schema(ontology_schema)
-          end
-        end
-
-      end
-
       context 'without symbols' do
         before do
           get :show,
